Create page 6-7
creates rectangles and elipses on opposite ends of the sketch that are filled with a random colour. moves one on the x and y axis to create the next rectangle and elipse filled with a different random colour. black vertical lines are created when the mouse is moved from left to right.